Moved more tx.rs functions into shared.
